Initial implementation of SimDevice support (#11)
* Initial implementation of SimDevice support

This commit leverages the SimDevices event in node-wpilib-ws to provide
Robot level SimDevice support. This allows for `WPILibWSRobotBase`
subclasses to register SimDevices on themselves, and have the Endpoint
automatically query and set values as appropriate.

Each `SimDevice` consists of fields which will be registered with the
`SimDevice` at creation time. These fields specify the data that the
`SimDevice` will handle. These will also get mapped to the actual
strings that are used in the WebSocket protocol to ensure that data gets
to and from robot code.

* Rename SimDevice value property

* Limit network writes to BIDIR/INPUT_TO_ROBOT_CODE